config: Extract to own file.
