+++ src/lib.rs
@@ -229,6 +229,7 @@ extern crate rand;
 /// documented by field.
 pub mod eraftpb;
 mod errors;
+mod config;
 mod log_unstable;
 mod progress;
 mod raft;
@@ -243,8 +244,9 @@ pub use self::errors::{Error, Result, StorageError};
 pub use self::log_unstable::Unstable;
 pub use self::progress::{Inflights, Progress, ProgressSet, ProgressState};
 pub use self::raft::{
-    quorum, vote_resp_msg_type, Config, Raft, SoftState, StateRole, INVALID_ID, INVALID_INDEX,
+    quorum, vote_resp_msg_type, Raft, SoftState, StateRole, INVALID_ID, INVALID_INDEX,
 };
+pub use self::config::Config;
 pub use self::raft_log::{RaftLog, NO_LIMIT};
 pub use self::raw_node::{is_empty_snap, Peer, RawNode, Ready, SnapshotStatus};
 pub use self::read_only::{ReadOnlyOption, ReadState};
@@ -269,7 +271,8 @@ pub mod prelude {
         Snapshot, SnapshotMetadata,
     };
 
-    pub use raft::{Config, Raft};
+    pub use raft::Raft;
+    pub use config::Config;
 
     pub use storage::{RaftState, Storage};
